+++ packages/core/src/Extension.ts
@@ -3,6 +3,7 @@ import { Command as ProseMirrorCommand } from 'prosemirror-commands'
 import { InputRule } from 'prosemirror-inputrules'
 import { Editor } from './Editor'
 import { Node } from './Node'
+import { Mark } from './Mark'
 import mergeDeep from './utilities/mergeDeep'
 import { GlobalAttributes, RawCommands, ParentConfig } from './types'
 import { ExtensionConfig } from '.'
@@ -31,7 +32,7 @@ declare module '@tiptap/core' {
      */
     addGlobalAttributes?: (this: {
       options: Options,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['addGlobalAttributes'],
     }) => GlobalAttributes | {},
 
     /**
@@ -40,7 +41,7 @@ declare module '@tiptap/core' {
     addCommands?: (this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['addCommands'],
     }) => Partial<RawCommands>,
 
     /**
@@ -49,7 +50,7 @@ declare module '@tiptap/core' {
     addKeyboardShortcuts?: (this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['addKeyboardShortcuts'],
     }) => {
       [key: string]: ProseMirrorCommand,
     },
@@ -60,7 +61,7 @@ declare module '@tiptap/core' {
     addInputRules?: (this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['addInputRules'],
     }) => InputRule[],
 
     /**
@@ -69,7 +70,7 @@ declare module '@tiptap/core' {
     addPasteRules?: (this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['addPasteRules'],
     }) => Plugin[],
 
     /**
@@ -78,7 +79,7 @@ declare module '@tiptap/core' {
     addProseMirrorPlugins?: (this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['addProseMirrorPlugins'],
     }) => Plugin[],
 
     /**
@@ -87,7 +88,7 @@ declare module '@tiptap/core' {
     extendNodeSchema?: ((
       this: {
         options: Options,
-        parentConfig: ParentConfig<ExtensionConfig<Options>>,
+        parent: ParentConfig<ExtensionConfig<Options>>['extendNodeSchema'],
       },
       extension: Node,
     ) => {
@@ -100,9 +101,9 @@ declare module '@tiptap/core' {
     extendMarkSchema?: ((
       this: {
         options: Options,
-        parentConfig: ParentConfig<ExtensionConfig<Options>>,
+        parent: ParentConfig<ExtensionConfig<Options>>['extendMarkSchema'],
       },
-      extension: Node,
+      extension: Mark,
     ) => {
       [key: string]: any,
     }) | null,
@@ -113,7 +114,7 @@ declare module '@tiptap/core' {
      onBeforeCreate?: ((this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['onBeforeCreate'],
     }) => void) | null,
 
     /**
@@ -122,7 +123,7 @@ declare module '@tiptap/core' {
     onCreate?: ((this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['onCreate'],
     }) => void) | null,
 
     /**
@@ -131,7 +132,7 @@ declare module '@tiptap/core' {
     onUpdate?: ((this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['onUpdate'],
     }) => void) | null,
 
     /**
@@ -140,7 +141,7 @@ declare module '@tiptap/core' {
     onSelectionUpdate?: ((this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['onSelectionUpdate'],
     }) => void) | null,
 
     /**
@@ -150,7 +151,7 @@ declare module '@tiptap/core' {
       this: {
         options: Options,
         editor: Editor,
-        parentConfig: ParentConfig<ExtensionConfig<Options>>,
+        parent: ParentConfig<ExtensionConfig<Options>>['onTransaction'],
       },
       props: {
         transaction: Transaction,
@@ -164,7 +165,7 @@ declare module '@tiptap/core' {
       this: {
         options: Options,
         editor: Editor,
-        parentConfig: ParentConfig<ExtensionConfig<Options>>,
+        parent: ParentConfig<ExtensionConfig<Options>>['onFocus'],
       },
       props: {
         event: FocusEvent,
@@ -178,7 +179,7 @@ declare module '@tiptap/core' {
       this: {
         options: Options,
         editor: Editor,
-        parentConfig: ParentConfig<ExtensionConfig<Options>>,
+        parent: ParentConfig<ExtensionConfig<Options>>['onBlur'],
       },
       props: {
         event: FocusEvent,
@@ -191,7 +192,7 @@ declare module '@tiptap/core' {
     onDestroy?: ((this: {
       options: Options,
       editor: Editor,
-      parentConfig: ParentConfig<ExtensionConfig<Options>>,
+      parent: ParentConfig<ExtensionConfig<Options>>['onDestroy'],
     }) => void) | null,
   }
 }
@@ -199,48 +200,55 @@ declare module '@tiptap/core' {
 export class Extension<Options = any> {
   type = 'extension'
 
+  name = 'extension'
+
+  parent: Extension | null = null
+
+  child: Extension | null = null
+
+  options: Options
+
   config: ExtensionConfig = {
-    name: 'extension',
+    name: this.name,
     priority: 100,
     defaultOptions: {},
   }
 
-  parentConfig: Partial<ExtensionConfig> = {}
-
-  options!: Options
-
-  constructor(config: ExtensionConfig<Options>) {
+  constructor(config: Partial<ExtensionConfig<Options>> = {}) {
     this.config = {
       ...this.config,
       ...config,
     }
 
+    this.name = this.config.name
     this.options = this.config.defaultOptions
   }
 
-  static create<O>(config: ExtensionConfig<O>) {
+  static create<O>(config: Partial<ExtensionConfig<O>> = {}) {
     return new Extension<O>(config)
   }
 
   configure(options: Partial<Options> = {}) {
-    return Extension
-      .create<Options>(this.config as ExtensionConfig<Options>)
-      .#configure(options)
-  }
-
-  #configure = (options: Partial<Options>) => {
-    this.options = mergeDeep(this.config.defaultOptions, options) as Options
+    this.options = mergeDeep(this.options, options) as Options
 
     return this
   }
 
-  extend<ExtendedOptions = Options>(extendedConfig: Partial<ExtensionConfig<ExtendedOptions>>) {
-    const extension = new Extension<ExtendedOptions>({
-      ...this.config,
-      ...extendedConfig,
-    } as ExtensionConfig<ExtendedOptions>)
+  extend<ExtendedOptions = Options>(extendedConfig: Partial<ExtensionConfig<ExtendedOptions>> = {}) {
+    const extension = new Extension<ExtendedOptions>(extendedConfig)
 
-    extension.parentConfig = this.config
+    extension.parent = this
+
+    this.child = extension
+
+    extension.name = extendedConfig.name
+      ? extendedConfig.name
+      : this.name
+
+    extension.options = {
+      ...extension.parent.options,
+      ...extension.options,
+    }
 
     return extension
   }