+++ packages/core/src/Mark.ts
@@ -14,6 +14,7 @@ import {
   GlobalAttributes,
   ParentConfig,
 } from './types'
+import { Node } from './Node'
 import { MarkConfig } from '.'
 import { Editor } from './Editor'
 
@@ -41,7 +42,7 @@ declare module '@tiptap/core' {
      */
     addGlobalAttributes?: (this: {
       options: Options,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['addGlobalAttributes'],
     }) => GlobalAttributes | {},
 
     /**
@@ -51,7 +52,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['addCommands'],
     }) => Partial<RawCommands>,
 
     /**
@@ -61,7 +62,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['addKeyboardShortcuts'],
     }) => {
       [key: string]: ProseMirrorCommand,
     },
@@ -73,7 +74,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['addInputRules'],
     }) => InputRule[],
 
     /**
@@ -83,7 +84,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['addPasteRules'],
     }) => Plugin[],
 
     /**
@@ -93,7 +94,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['addProseMirrorPlugins'],
     }) => Plugin[],
 
     /**
@@ -102,7 +103,7 @@ declare module '@tiptap/core' {
     extendNodeSchema?: ((
       this: {
         options: Options,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['extendNodeSchema'],
       },
       extension: Node,
     ) => {
@@ -115,9 +116,9 @@ declare module '@tiptap/core' {
     extendMarkSchema?: ((
       this: {
         options: Options,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['extendMarkSchema'],
       },
-      extension: Node,
+      extension: Mark,
     ) => {
       [key: string]: any,
     }) | null,
@@ -125,11 +126,11 @@ declare module '@tiptap/core' {
     /**
      * The editor is not ready yet.
      */
-     onBeforeCreate?: ((this: {
+    onBeforeCreate?: ((this: {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['onBeforeCreate'],
     }) => void) | null,
 
     /**
@@ -139,7 +140,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['onCreate'],
     }) => void) | null,
 
     /**
@@ -149,7 +150,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['onUpdate'],
     }) => void) | null,
 
     /**
@@ -159,7 +160,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['onSelectionUpdate'],
     }) => void) | null,
 
     /**
@@ -170,7 +171,7 @@ declare module '@tiptap/core' {
         options: Options,
         editor: Editor,
         type: MarkType,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['onTransaction'],
       },
       props: {
         transaction: Transaction,
@@ -185,7 +186,7 @@ declare module '@tiptap/core' {
         options: Options,
         editor: Editor,
         type: MarkType,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['onFocus'],
       },
       props: {
         event: FocusEvent,
@@ -200,7 +201,7 @@ declare module '@tiptap/core' {
         options: Options,
         editor: Editor,
         type: MarkType,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['onBlur'],
       },
       props: {
         event: FocusEvent,
@@ -214,7 +215,7 @@ declare module '@tiptap/core' {
       options: Options,
       editor: Editor,
       type: MarkType,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['onDestroy'],
     }) => void) | null,
 
     /**
@@ -227,7 +228,7 @@ declare module '@tiptap/core' {
      */
     inclusive?: MarkSpec['inclusive'] | ((this: {
       options: Options,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['inclusive'],
     }) => MarkSpec['inclusive']),
 
     /**
@@ -235,7 +236,7 @@ declare module '@tiptap/core' {
      */
     excludes?: MarkSpec['excludes'] | ((this: {
       options: Options,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['excludes'],
     }) => MarkSpec['excludes']),
 
     /**
@@ -243,7 +244,7 @@ declare module '@tiptap/core' {
      */
     group?: MarkSpec['group'] | ((this: {
       options: Options,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['group'],
     }) => MarkSpec['group']),
 
     /**
@@ -251,7 +252,7 @@ declare module '@tiptap/core' {
      */
     spanning?: MarkSpec['spanning'] | ((this: {
       options: Options,
-      parentConfig: ParentConfig<MarkConfig<Options>>,
+      parent: ParentConfig<MarkConfig<Options>>['spanning'],
     }) => MarkSpec['spanning']),
 
     /**
@@ -260,7 +261,7 @@ declare module '@tiptap/core' {
     parseHTML?: (
       this: {
         options: Options,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['parseHTML'],
       },
     ) => MarkSpec['parseDOM'],
 
@@ -270,12 +271,12 @@ declare module '@tiptap/core' {
     renderHTML?: ((
       this: {
         options: Options,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['renderHTML'],
       },
       props: {
         mark: ProseMirrorMark,
         HTMLAttributes: { [key: string]: any },
-      }
+      },
     ) => DOMOutputSpec) | null,
 
     /**
@@ -284,7 +285,7 @@ declare module '@tiptap/core' {
     addAttributes?: (
       this: {
         options: Options,
-        parentConfig: ParentConfig<MarkConfig<Options>>,
+        parent: ParentConfig<MarkConfig<Options>>['addAttributes'],
       },
     ) => Attributes | {},
   }
@@ -293,48 +294,55 @@ declare module '@tiptap/core' {
 export class Mark<Options = any> {
   type = 'mark'
 
+  name = 'mark'
+
+  parent: Mark | null = null
+
+  child: Mark | null = null
+
+  options: Options
+
   config: MarkConfig = {
-    name: 'mark',
+    name: this.name,
     priority: 100,
     defaultOptions: {},
   }
 
-  parentConfig: Partial<MarkConfig> = {}
-
-  options!: Options
-
-  constructor(config: MarkConfig<Options>) {
+  constructor(config: Partial<MarkConfig<Options>> = {}) {
     this.config = {
       ...this.config,
       ...config,
     }
 
+    this.name = this.config.name
     this.options = this.config.defaultOptions
   }
 
-  static create<O>(config: MarkConfig<O>) {
+  static create<O>(config: Partial<MarkConfig<O>> = {}) {
     return new Mark<O>(config)
   }
 
   configure(options: Partial<Options> = {}) {
-    return Mark
-      .create<Options>(this.config as MarkConfig<Options>)
-      .#configure(options)
-  }
-
-  #configure = (options: Partial<Options>) => {
-    this.options = mergeDeep(this.config.defaultOptions, options) as Options
+    this.options = mergeDeep(this.options, options) as Options
 
     return this
   }
 
-  extend<ExtendedOptions = Options>(extendedConfig: Partial<MarkConfig<ExtendedOptions>>) {
-    const extension = new Mark<ExtendedOptions>({
-      ...this.config,
-      ...extendedConfig,
-    } as MarkConfig<ExtendedOptions>)
+  extend<ExtendedOptions = Options>(extendedConfig: Partial<MarkConfig<ExtendedOptions>> = {}) {
+    const extension = new Mark<ExtendedOptions>(extendedConfig)
 
-    extension.parentConfig = this.config
+    extension.parent = this
+
+    this.child = extension
+
+    extension.name = extendedConfig.name
+      ? extendedConfig.name
+      : this.name
+
+    extension.options = {
+      ...extension.parent.options,
+      ...extension.options,
+    }
 
     return extension
   }