+++ packages/core/src/helpers/getAttributesFromExtensions.ts
@@ -1,12 +1,14 @@
-import createExtensionContext from './createExtensionContext'
 import splitExtensions from './splitExtensions'
+import getExtensionField from './getExtensionField'
 import {
   Extensions,
   GlobalAttributes,
   Attributes,
   Attribute,
   ExtensionAttribute,
+  AnyConfig,
 } from '../types'
+import { NodeConfig, MarkConfig } from '..'
 
 /**
  * Get a list of all extension attributes defined in `addAttribute` and `addGlobalAttribute`.
@@ -25,15 +27,22 @@ export default function getAttributesFromExtensions(extensions: Extensions): Ext
   }
 
   extensions.forEach(extension => {
-    const context = createExtensionContext(extension, {
+    const context = {
       options: extension.options,
-    })
+    }
 
-    if (!extension.config.addGlobalAttributes) {
+    const addGlobalAttributes = getExtensionField<AnyConfig['addGlobalAttributes']>(
+      extension,
+      'addGlobalAttributes',
+      context,
+    )
+
+    if (!addGlobalAttributes) {
       return
     }
 
-    const globalAttributes = extension.config.addGlobalAttributes.bind(context)() as GlobalAttributes
+    // TODO: remove `as GlobalAttributes`
+    const globalAttributes = addGlobalAttributes() as GlobalAttributes
 
     globalAttributes.forEach(globalAttribute => {
       globalAttribute.types.forEach(type => {
@@ -54,21 +63,28 @@ export default function getAttributesFromExtensions(extensions: Extensions): Ext
   })
 
   nodeAndMarkExtensions.forEach(extension => {
-    const context = createExtensionContext(extension, {
+    const context = {
       options: extension.options,
-    })
+    }
 
-    if (!extension.config.addAttributes) {
+    const addAttributes = getExtensionField<NodeConfig['addAttributes'] | MarkConfig['addAttributes']>(
+      extension,
+      'addAttributes',
+      context,
+    )
+
+    if (!addAttributes) {
       return
     }
 
-    const attributes = extension.config.addAttributes.bind(context)() as Attributes
+    // TODO: remove `as Attributes`
+    const attributes = addAttributes() as Attributes
 
     Object
       .entries(attributes)
       .forEach(([name, attribute]) => {
         extensionAttributes.push({
-          type: extension.config.name,
+          type: extension.name,
           name,
           attribute: {
             ...defaultAttribute,