+++ packages/core/src/index.ts
@@ -12,7 +12,8 @@ export { default as markPasteRule } from './pasteRules/markPasteRule'
 export { default as callOrReturn } from './utilities/callOrReturn'
 export { default as mergeAttributes } from './utilities/mergeAttributes'
 
-export { default as createExtensionContext } from './helpers/createExtensionContext'
+export { default as coordsAtPos } from './helpers/coordsAtPos'
+export { default as getExtensionField } from './helpers/getExtensionField'
 export { default as findChildren } from './helpers/findChildren'
 export { default as findParentNode } from './helpers/findParentNode'
 export { default as findParentNodeClosestToPos } from './helpers/findParentNodeClosestToPos'
@@ -26,6 +27,7 @@ export { default as isNodeActive } from './helpers/isNodeActive'
 export { default as isNodeEmpty } from './helpers/isNodeEmpty'
 export { default as isNodeSelection } from './helpers/isNodeSelection'
 export { default as isTextSelection } from './helpers/isTextSelection'
+export { default as posToDOMRect } from './helpers/posToDOMRect'
 
 export interface Commands {}