+++ packages/extension-code-block-lowlight/src/code-block-lowlight.ts
@@ -8,14 +8,13 @@ export interface CodeBlockLowlightOptions extends CodeBlockOptions {
 
 export const CodeBlockLowlight = CodeBlock.extend<CodeBlockLowlightOptions>({
   defaultOptions: {
-    ...CodeBlock.config.defaultOptions,
+    ...CodeBlock.options,
     lowlight,
   },
 
   addProseMirrorPlugins() {
     return [
-      // disable for now, see: https://github.com/ueberdosis/tiptap-next/issues/259#issuecomment-817954835
-      // ...this.parentConfig.addProseMirrorPlugins?.() || [],
+      ...this.parent?.() || [],
       LowlightPlugin({
         name: 'codeBlock',
         lowlight: this.options.lowlight,