+++ packages/extension-horizontal-rule/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.4](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-horizontal-rule@2.0.0-beta.3...@tiptap/extension-horizontal-rule@2.0.0-beta.4) (2021-04-16)
+
+**Note:** Version bump only for package @tiptap/extension-horizontal-rule
+
+
+
+
+
+# [2.0.0-beta.3](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-horizontal-rule@2.0.0-beta.2...@tiptap/extension-horizontal-rule@2.0.0-beta.3) (2021-04-15)
+
+**Note:** Version bump only for package @tiptap/extension-horizontal-rule
+
+
+
+
+
 # [2.0.0-beta.2](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-horizontal-rule@2.0.0-beta.1...@tiptap/extension-horizontal-rule@2.0.0-beta.2) (2021-04-07)