+++ packages/extension-italic/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.3](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-italic@2.0.0-beta.2...@tiptap/extension-italic@2.0.0-beta.3) (2021-04-16)
+
+**Note:** Version bump only for package @tiptap/extension-italic
+
+
+
+
+
+# [2.0.0-beta.2](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-italic@2.0.0-beta.1...@tiptap/extension-italic@2.0.0-beta.2) (2021-04-15)
+
+**Note:** Version bump only for package @tiptap/extension-italic
+
+
+
+
+
 # [2.0.0-beta.1](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-italic@2.0.0-alpha.11...@tiptap/extension-italic@2.0.0-beta.1) (2021-03-05)
 
 **Note:** Version bump only for package @tiptap/extension-italic